refactor: Added useGroupMutate hook
